bev_centerline/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bev-centerline-mm"
version = "0.1.0"
description = "Dual-vision Qwen3.5 + DINOv3 project for BEV road centerline JSON generation."
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "transformers>=4.57.0",
  "accelerate>=0.30",
  "peft>=0.11",
  "pillow>=10.0",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "numpy>=1.24",
  "matplotlib>=3.7",
]

[tool.setuptools.packages.find]
where = ["src"]

